requirements.txt
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.2
botocore==1.31.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
easyocr==1.7.0
filelock==3.12.2
Flask==2.2.5
Flask-Cors==4.0.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-vision==3.4.4
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==20.1.0
idna==3.4
imageio==2.31.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy_loader==0.3
llvmlite==0.41.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numba==0.58.0
numpy==1.25.1
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.24.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycryptodomex==3.19.0
PyJWT==2.7.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
scikit-image==0.21.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
sympy==1.12
tifffile==2023.7.18
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.2.3
